Display coverage information for backlog items
Go to test plan view of a milestone. Each backlog item can now:
* Display number of tests that covers it
* Display the status of the tests as an aggregation of the test results.

Test definitions are now automatically loaded on first run. When they
are being loaded, skeletons are displayed in lieu of the nb tests &
icon.

Implementation details
----------------------

FTR we took the decision to not follow the acceptance criteria
recommendation (define a new route to retrieve the coverage
information).

This is due to our lack of test plan volumetry. We don't know how the
computation of coverage information is costly in real life environment.

We identified 3 ways to get this information:
A. Via the backlog items: milestone/:id/testplan.
B. Via a new dedicated route (A.C recommendation): backlog_items/:id?????
C. Via an test definitions of a backlog_items:
backlog_items/:id/test_definitions

If we encounter a performance penalty by augmenting A., we are doomed
because we won't be able to change the route without introducing a
breaking change.

The information we need is already available in the existing route which
retrieves tests under a backlog item. Loading extra information (test
names & co) while the backlog item is collapsed is considered acceptable
as far as we know. If reality shows that we are wrong and we need a
dedicated route, then we will be able to change our mind and to
introduce it without BC.

Part of story #14887: Display Backlog Items status

Change-Id: I96f9e55ee23b7fc1ea91f92d947a718df9aa99af