Remove dependency to perl-CGI
The "SVN core" repositories have a commit hook depending on perl-CGI
(commit-email.pl ➡ svn-checkins.pl ➡ session.pl) but this dependency is
not explicitely required by the package tuleap-core-subversion on
CentOS/RHEL 7. This means the post commit hook is broken on instance
that uses "SVN core" repositories expect if the dependency is installed
for other reasons (tuleap-core-cvs requires it).

However, the perl-CGI is not needed anymore, the code does not make any
use of it so instead of fixing the packaging this contribution removes
completely the dependency.

Closes request #15046: Missing dependency to run post commit hook of the "SVN core" repositories

Change-Id: Ida00953455dba1ecd9915a2af81d4d6082527034