Stop loading the database.inc into the global scope
The usage of require_once with a path that cannot be resolved during a
static analysis does not play well with the taint analysis feature of
Psalm [0]. It's also a good opportunity to do a much needed cleanup.

This contribution only deals with the database.inc. Another contribution
will deal with the local.inc.

Part request #15015: Initiate usage of Psalm taint analysis feature

[0] https://github.com/vimeo/psalm/issues/3691

Change-Id: If6c13f3ad330611eb3157e84aba532e67555b240