request #15065: Changing password should revoke all OAuth2 tokens
If a user's account is compromised by an attacker, the attacker could try to persist
some of its accesses by registering an evil OAuth2 app into the account. By revoking
all the OAuth2 tokens and codes when users change their passwords, we make sure
the users have purposely wanted to share accesses with the app at least once.

All the OAuth2 access tokens, refresh tokens and authorization codes of
a user are removed when the password of the account is changed.
Authorizations are kept to not make the process too annoying for the
user.

Change-Id: I20448a793fb894e6fd17ec3b656842a985502cc5