request #15072: Bump to sass-loader 9.0.1 and move to Dart Sass
The changelog of sass-loader 9.0.0 strongly recommends to migrate to
Dart Sass so this contribution also deal with that.
fibers has not been installed to avoid dealing with node-gyp ("Why do I
need g++ to built my stylesheet?"), in the context of Tuleap with the
highly concurrent build process it does not seem to improve performance
a lot.

Changelog:
https://github.com/webpack-contrib/sass-loader/releases/tag/v9.0.0
https://github.com/webpack-contrib/sass-loader/releases/tag/v9.0.1

Change-Id: I005fcea3d9dc510fcadef9d1ba0f0c87136724d1