request #15074: Artifact deletions are never processed when a Redis instance is available but there are no async workers
Messages to process the gets added to the queue but since there are no worker they never
get processed. The deleted artifacts never gets archived and the multitudes of dependencies
(permissions, cross refs, unsubscribe of notifications...) are never cleaned.

To reproduce you need to have a working Redis known by Tuleap (for
example to store metrics) but the sys_nb_backend_workers setting set to 0.

Change-Id: Ia4dcea11404fae718bc37042cac0938255465ad2