Cache users
part of: story #15066 Map Jira users with Tuleap users

The goal of this commit is to provide a cache storing a collection of users
indexed by their Jira account ids. It will avoid us to query systematically
UserManager to identify users we have already met.

Behavior:

When a user email address matches with a Tuleap user, we store it in cache.
When it does not match, or the user does not share email address, we store forge_user_instead.

-- No functional change expected --

Change-Id: I35443432092524fb4c1f91bc41fbe19b21988a0b