Stop loading the local.inc into the global scope
The usage of require_once with a path that cannot be resolved during a
static analysis does not play well with the taint analysis feature of
Psalm [0]. It's also a good opportunity to do a much needed cleanup.
As it allows Psalm to better analyze the code a lot of new errors are
found. The baseline has been updated to integrated them.

Administrators having customized site-content files will need to check
and update their files if a configuration setting value is accessed via
$GLOBALS.

Part request #15015: Initiate usage of Psalm taint analysis feature

[0] https://github.com/vimeo/psalm/issues/3691

Change-Id: Icc3c0edaa029051ea85641b9474c2e5623d8e519