request #15079 Correct cypress workflow tests
This happens because we are in a strange case:
tlp dropdown cause a flickering on tabs when the screen size is at a certain size.
When cypress get the button, it also get its x/y coordinates.
When it tries to click on it, the flickering cause the button to not be in the same place.

We try to update css properties to make it correct, but we did not reach a point
where we never have the flickering, and the issue is present depending on the size screen.
That's why we fix a viewport of screen for the test.

Change-Id: Ie646377081e09dc4641ddf0d9ba2c947b24e67ed