Remove taint analysis false positives in TLP doc
The taint analysis detects some LFI and XSS in the TLP doc because the
code does not make it easy to actually detect the `$current_section`
variable is built from known values. This small code changes make it
easier for the taint analysis engine to detect that.

Note that you may need to comment the line `require($local_inc)` in
pre.php to see the issue.

Part request #15015: Initiate usage of Psalm taint analysis feature

Change-Id: Iacaed6ce6862d85dd9b90d1614bb3bb7c88ce50e