REST representations of tracker are immutables
The tracker plugin only uses immutable representations for all its REST
endpoints.

The tracker plugin is the first plugin to be converted in reason of its
central role. As the immutability of a class infects all the other
classes that compose it, it's not easy to convert a part of the API and
not end up converting the tracker plugin. Also, it shows it is doable
everywhere since the tracker plugin comes with a lot of baggage that does
not make the conversion straightforward: big class hierarchy with child
class updating parent properties, local caches everywhere, no clean
separation between the infrastructure and logic layers…

Part of request #15111: REST representations should not be mutable

Change-Id: I8ab00ab8728548576e8f0e8e4060065583282641