Tracker admin can define a color for none value
part of story #14824 change color of "none" value

How to test:
 - have a selectbox defined as required
   => you can't save a color for none value

 - have a selectbox defined as not required
  => you can select a color (legacy or tlp) for none value

If you check "required" and define a color for none, the color won't be saved.
If you already have a color for none value and you check "required", the color associated to none will be erased in db.

I'll deal with UI (adding disable property and title) for color selection in a dedicated patch.

Change-Id: I12d48cda1abadbc60863dcb73baafa4dda4c4464