request #15116: Remove the Tuleap tours
They only works on FP pages which are becoming a rare species nowadays.
The 2 tours that are provided by Tuleap are no more usable. The feature
is dead code for the vast majority of usages as the customization of the
tours was never documented and nobody complained of their absences with
the migration to BurningParrot.

Change-Id: I8ee9c808a1c961a7d212e7d312af9ae1a536319c