Use Git 2.18 on CentOS/RHEL 7
Access to the package sclo-git212-git on RHEL 7 is not easy as it is not
part of an "official" SCL. The installation guide does not take this
into account because there is no official way to install the SCLO
package.

As the rh-git218-git package is not available on CentOS/RHEL 6, Git 2.12
is kept on these instances.

As we need a way to deploy the required change to the gitolite3
configuration automatically (it is a requirement for the container-based
instance), the configuration is now deployed by the 'tuleap-cfg
site-deploy' command.
To not deal with anothe switch when generating the gitolite3
configuration, all the instances are migrated to the Tuleap SSH key management.
The feature can be considered stable and well-tested in real life as it has been
here for the past 3 years and there is no reason to keep supporting both ways.
Removal of the dedicated migration panel in the site administration will come in
a dedicated contribution.

To test, do not forget to install the 'rh-git218-git' package and to
redeploy the gitolite3 configurations with
'tuleap-cfg site-deploy:gitolite3-config'.

Part of request #14963: tuleap-git-plugin does not install on RHEL7

Change-Id: I6837b19fc155b26c5d3007f4f9fa8d599bff9384