Map users in userpicker fields
part of: story #15066 Map Jira users with Tuleap user

/!\ PLEASE READ BEFORE TESTING

The current behavior of the mapping is as following:
When the user is found on Tuleap -> It is bound to the selectbox
When not found, no changelog or value can be imported because the
importer does not allow Tracker_Importer user to be bound to a
selectbox field. Since it is not a real user, it is not shocking to
not have a value in this case.

How to test:
- Import a Jira tracker containing issues on which your are assignee or
  reporter

- Once imported, open the different artifacts
--> You are assigned to artifacts were you are supposed to be assigned
--> You are reporter of issues you have reported
--> Assignations history is imported

Change-Id: I06fa8c0d284c43b37e523233ad88d5283a8dbc46