request #15121: Extract FilterInlineTable out of TLP
Its usage does not justify adding it to every BurningParrot page ever
(which is the case for TLP). That said, the size gain is still marginal
for the TLP asset (about 4 KiB).
Now that we have webpack, we can safely import it only where it is used.

I took this opportunity to export constants for CSS class names.

How to test:
- No functional change. Pages using it still function and filter tables
correctly. Usage is mainly in site administration and project
administration.

Change-Id: Ief3d65773af0bb3ddc5e96a015969fa46fc11221