Detect the manual merge of a PR before updating the merge statuses
Marking the PRs that have been manually merged before doing the
detection of the merge statuses can save resources: no need to compute
the merge status of a PR that have been manually merged.
It also make the code more resilient: in case of an issue when computing
the merge statuses (e.g. request #15077) at least the already merged PRs are
closed which does not completly broke the feature.

Closes request #15082: FastForward not detected when checking mergeability

Change-Id: I2fc615eee667f79129f566a8724d9021fe772b38