Display relative date according to user preference
Go to an artifact and look at the follow-up comments' dates. The date
displayed depends on the user preference:

* **absolute_first-relative_tooltip**: Absolute date (relative is in the :hover)
* **relative_first-absolute_tooltip**: Relative date (absolute is in the :hover)
* **absolute_first-relative_shown**: Absolute first (relative is in small on top or next when inline)
* **relative_first-absolute_shown**: Relative first (absolute is in small on top or next when inline)

If no pref or user is anonymous, then the current display is used:
Relative date (absolute is in the :hover).

**Note:** this it is not possible to define the preference through the
ui yet, the reviewer must insert the preference through REST API or
directly in the db (preference_name = relative_dates_display).

**Note2:** tlp style has been copied as is in flaming parrot theme, so
that we don't have to clutter a tlp component with flaming parrot
specific class names.

Part of story #14828: display dates with both absolute and relative dates

Change-Id: Ice262d7587ce178da31c9ae65a360b9df243856f