request #15124: Failures of the PHP unit tests suite due to an incorrect casing of the \UserXMLExporter class
The tests in the \Tracker_FormElement_Field_List_Bind_UgroupsTest generate mocks
based on the \UserXMLExporter class, however they use an incorrect case.
As a result the class is not loaded (Composer autoloader is case sensitive) and the
class end up being replaced by the mock framework. The other test classes using
\UserXMLExporter might fail if they are executed after this one since they do not
have accessto the real class anymore, PHP is not case sensitive with the class names
so the autoloading is not triggered again.

Change-Id: I5f7796d25b4df223cb32a9f80b8c3b4222fe653c