request #15125: Bump lodash to 4.17.49
A low level security issue has been identified:
https://github.com/advisories/GHSA-p6mc-m468-83gw

There is little to no risk in the context of Tuleap,
to be exploitable is some ways it would require to zip objects
with properties defined by users.

Upgrade has been applied wherever it was possible while respecting the
existing semver constraint.

Change-Id: I465a02e9cb3277d6ef4937ef92a7f072827d7315