User can change relative dates display preference
In user preferences » appareance & language, just like Username display,
we can now choose the way relative dates are displayed.

In order to iterate fast, the chosen strings for the select are
considered good enough for a first step (it mimics Username ones). A
dedicated changeset will deal with user experience later, with the help
of our Design team.

The change is reflected in artifact follow-up comments.

Part of story #14828: display dates with both absolute and relative dates

Change-Id: Ic5cb4002e42e4806d87f7a035a3a575862444a24