Use tlp-relative-date in legacy docman
Change your user preference "relative_dates_display" to
"absolute_first-relative_shown". (You have to do it manually in the db)

Go to history of a file, embedded file or link. You should see the date
in the form YYYY-MM-DD with the relative version (X days ago) on top.

Part of story #14828: display dates with both absolute and relative dates

Change-Id: I58d5184156431dde616e7867edf400046c315d43