Remove not needed method \Tuleap\Dao\UserSuspensionDao::verifySuspension
This method is only called after we call
UserSuspensionDao::suspendAccount() which makes it useless. Either the
query in suspendUser has crashed in which case we will never reach this
part of the code or if the query has succedded the only way for this
method to return true is a race condition (if we exclude DB issues).

Part of story #13811: Have a customisable reactivation period for the account

Change-Id: I884d21a6b7667de9c78e4d1015d9a9bf3712a6b3