request #16137: Some E2E tests with experimentalFetchPolyfill option enabled fail
Tests has been rewritten to not need a fetch polyfill to instrument the
queries. It also decorrelates the E2E tests from the implementation
details since the main thing we want to check in this situation is what
the users see, not what requests they sent.

Replayability of the tracker_artifact.spec.ts has also been improved a
bit to keep some sanity while debugging the issue.

Change-Id: I9c3edf8eea49397e657c668d5c0243dc4ddbcc41