request #16575: A user can decide to receive notifications for every status change
The custom notification settings 'Notify me on status change' has a
different behavior than the others. When a user chooses 'Notify me when artifacts are created'
or 'Notify me on every change', notifications will be sent even if the
user was not part of the global notification settings (in fact the user
gets added to it). With the 'Notify me on status change' mode if the
user is not somehow in the global notification settings no notifications
will be sent.
This contribution aligns the behaviors between the different modes and
the user is now added to the 'Global notification settings' like for the
other modes.

Change-Id: I188a22f14d260e75dbff4d73a22930d444eaabe8