Introduce a way to find the name of the browsers used to access a Tuleap instance
Only the names of browsers officially supported (or that were supported)
by Tuleap are detected.
The detection is done with an approach similar to the one used by Matomo [0].
The detected browser information are stored in a class with a primitive
API but it should be enough to cover the needs in Tuleap. Pulling a
full-blown browser detection library such as piwik/device-detector comes
with a non negligible performance cost and offers far too much
information that Tuleap does not need at the moment.

The opportunity has been taken to extract the detection from the
\HTTPRequest class for a better separation of concerns.

This is a technical change to introduce a new internal API, no
functional change is expected.

Part of request #16147: Add a label on the web hits metric to determine which browser was used

[0] https://github.com/matomo-org/device-detector

Change-Id: I567fc5ce141e6df51a96e53532c182067a680bb7