Add 'browser' label to the http_responses_total metric
This information can be retrieved by Prometheus via the /metrics
endpoint.

Reviewers wanting to check out the /metrics page needs to have
configured their Tuleap instance to have access to a Redis instance and
to have configured the prometheus_metrics plugin.

The tuleap/prometheus-client library has been bumped to 1.1.0 to include
a bugfix and gives us the opportunity to detect changes in the structure
of the metrics (e.g. the addition of a new label).

Closes request #16147: Add a label on the web hits metric to determine which browser was used

Change-Id: Ic9958d67fd301ad3111c2e7735289022b63e64a4