Not planned tests are displayed after the planned/executed ones
Outside of that, the order of the tests stay as it previously was.

A small change has been done on the backend in a attempt to pre-sort the
returned information to limit the work the front-end needs to do. It can
helps saving a few moves when sorting the test definitions array.

Part of story #15094: have more effective test plan

Change-Id: I9009c12dbd5a74b5008f04735282290969857220