Improve route milestones/$id/backlog to get all artifacts (closed and open)
This is part of request #16146 All artifacts linked to a release are not displayed in Project Milestones

In this patch, the route rest milestones/$id/backlog is improve with a query selector.
The query can only be '{"status":"all"}'.
If there is not query, there is the same comportment like before (only open artifact are returned).

How to test:
	- Have a milestone with some artifact in its backlog
	- Change status of some artifacts to done

Expected results:
	- When you use the route rest with the query
		-> All artifacts are returned

	- Without query
		-> Only open artifacts are returned

Change-Id: I0853ab2a0d56249f5554b4f1fb584d2a6c351856