Create the new backlog admin route
This is part of story #15122 [release management] management of Aggregator projects

A new URL has been created to go to the aggregator read only top
planning page.

If you are in the aggregator project and click to the edit button of the
top planning, you display this new page.

This page content is for now empty, only the breadcrumb is displayed.

Change-Id: Iaf0d93a4e2e86e92d4bad8a81ae8f717eb62cebe