request #16178 Project milestone should be on Scrum & Agile ALM default dashboard
Project Milestone widget can be imported via XML import.

The only configuration supported right now is a widget that displays the content
of the current project. As project XML import is 1 projet by 1 project it's the
expected behaviour anyway.

The Scrum and Agile ALM templates default dashboard are update to remove the
useless note at creation and exposes the new widget when plugin is activated.

In order to test, you will have to clear caches (events) as well as re-generate
templates with `make generate-templates` (or the docker variant if you don't have
xsltproc locally).

Change-Id: Ia65390bb3f890b17ab5e4c2ed82c893be4aa06ab