request #16175 Test environment can have data
Create and purge data directory in a test environment.

When using create_test_env to generate a new project with actual content,
this content might need to refer to assets in  directory.
This directory is not copied yet 😭.

In order to reduce the risk of a huge mistake the clean-up
part is not "generic" rm -rf. It only remove `data` directory
and the first level of files within.

Change-Id: If0ac800f5bc262c73ee569d119dedfaa53b24c4d