Display page content
This is part of story #15122 [release management] management of Aggregator projects

The contributor projects and their plannable items trackers are now
displayed in the Aggregator top planning read only page.

To test it, you have to uninstall and reinstall the plugin and then add
data by hand in the two plugin tables.

No custom CSS has been done for now. It will be done in a dedicate commit

Change-Id: I02ad13fb427dbbc0324eb4f28532578217f1f9cf