Do not allow contributor projects to change milestone tracker in planning
Part of story #15122 [release management] management of Aggregator
projects

How to test:
- After installing the plugin, edit manually your DB to configure
  aggregator and contributor projects.
- In a contributor project, define a new Tracker (for example
  "Products"). Set it on top of Release in the Tracker hierarchy.
- Go to Scrum administration and Edit the Release planning (assuming it
is the root planning). Change the "target" tracker from "Releases" to
"Products".
- On submit, the modification should NOT be done and you should see a
  feedback message telling you that the modification is not allowed.
- No functional change expected on other planning modifications.

Notes to reviewers:
- A next contribution will disallow this action in the
UI. To keep the review size smaller, it is not done on this
contribution. Similarly, on this contribution end-users have no
explanation why the planning can't be modified. It will be done in the
UI with a message.

Change-Id: I6ac240b5804f42dc0d6b39a0bdd6995417bf772a