x-pagination-size should display the number of available backlog items in milestones/$id/backlog with query
This is part of request #16146 All artifacts linked to a release are not displayed in Project Milestones

In route rest milestones/$id/backlog with the query '{"status":"all"}', the pagination-size is not good.

Currently it's displayed the number of returned artifacts instead of the number of available artifacts.

How to test:
	- Use route rest milestones/$id/backlog with the query '{"status":"all"}'

Expected results:
	- The pagination size is the number of available artifacts.

Change-Id: I493d854254fb00220dc833ce600af128e9f94c68