Extract a dedicated presenter & template for planning edition
Part of story #15122 [release management] management of Aggregator
projects

This is a refactoring to make it easier to change the presentation of
the planning edition. The multi_project_backlog plugin will need to
alter this presentation to disallow changing the milestone tracker. This
contribution decouples planning edition from planning creation, which
will make this easier.

How to test:
- No functional change expected in planning edition. Burnup and velocity
warnings should show as before. Cardwall configuration should work as
before.

Change-Id: I3732015535a2dad6f29c2040f4c8e9e27f71ac2a