Refresh the backlog items when creating a new campaign
As the creation of a campaign can change the status of the backlog items
(e.g. some non planned tests can become planned) the view of the backlog
items can become inconsistent.

The approach taken here is quite primitive: we reload all the backload
items from scratch (meaning we loose the information about their
expansion) whenever a campaign is successfully created. While
aggressive, it should be good enough as campaign creation is not that
frequent.

Part of story #15094: have more effective test plan

Change-Id: Ie5543d3e0580d32151a8e70d364b926f60fb9e46