Synchronize plannable item trackers after contributor plannig update
This is part of story #15122 [release management] management of Aggregator projects

When updating the backlog item trackers of a root planning of a
contributor project, the new selected trackers are updated in the SQL
sync table of the multi backlog project plugin.

This must be visible in root planning read only view in the aggregator project.

Change-Id: I5a4140f940a73671b267b09a06083f706e006d15