Extract update-specific request validator for planning admin
Part of story #15122 [release management] management of Aggregator
projects

This is a refactoring to make it easier to change the request validation
rules for root plannings. The existing RequestValidator is used for
planning creation from a form and XML and for planning update. Changing
it would bring unintended side-effects. Also, planning updates have
different validation constraints than planning creations.

How to test:
- No functional change expected.
- Name, "planning tracker" and backlog trackers are mandatory and cannot
be empty.

Change-Id: I20d5678e7e57362347891847bf95a7388bccf691