Add a minimal export report of a test plan
For now only the project milestone names are exported in a XSLX
document. It is not really helpful by itself, the primary goal of the
contribution is set up the architecture of the export generation.

The report is purposely built without any dependency to an external
library. The transformation into a XSLX document is done as the last
step of the chain in a isolated context. This approach ease the long
term maintainance (it becomes easier to follow the library API changes
at upgrades) and avoid relying on hard to crasp and very domain specific
API. Also, it can potentially ease the process if other export formats need
to be added in the future.

The code needed to generate the report is loaded asynchronously in
reason of its large size.

Part of story #15096: export a report of tests for a milestone

Change-Id: Ia7f02214a0e4cf96a3300299a5020a9fb82e7a32