Check contributor title semantics before aggregator milestone creation
This is part of story #16124 [release management] ensure Aggregator release creation is possible

How to test:
- Setup aggregator and contributor projects
- In a contributor project's top milestone tracker, unset the Title
  semantic.
- When you browse the aggregator project's top backlog planning, you
  should NOT see the "Create milestone" button.
- If you set back the Title semantic, the button should appear.

Note to reviewer:
I would have preferred not to use Exceptions, but in this case we will
want to log what kind of error happened for debug purposes. I cannot
simply return a null collection as this would lose some information on
the precise error encountered. Therefore I used Exceptions.

Change-Id: I89f7cf2fddb8f79ce7770e617e1a00f60321b1f2