Add description field and semantic in release tracker of scrum template
This is part of story #16124 [release management] ensure Aggregator release creation is possible

A text field named description is added to the scrum template in the
release tracker. This field is also added in the description semantic.

To test this patch, you will need to rebuild the templates with the
following command:

make -C plugins/agiledashboard/ generate-templates

Change-Id: I455c984d178cdb38803ea1507cef3a14cd197068