Decrease timeAgoInWordsUsage
part of story story #14828: display dates with both absolute and relative

How to test:
 - make a push in a git repository, go in dashboard My Git pushed
   => date ok

 - perform a SVN commit
   => in home page commit date is displayed ok

 - create news
   => date is ok

This is the last poart of TimeAgoInWords usage removal.

I did not update TV3 stuff, the goto reference seems also linked to TV3.
The widget MyRss will not be taken in account, I didn't find a way to do it

Change-Id: Ia7151e94665f416661c4235e524b11d5b0e7c1ea