Move "New project" in dedicated dropdown
If user can create new project, then they must now use a dedicated
dropdown in the navbar [+ New].

This is an introduction of the button/dropdown. For now this dropdown
only offers the possibility to create a new project. Next it will let
the user to create directly a bug or a story in a given project, but it
will be a dedicated changeset.

Part of story #16205: « + New » button

Change-Id: Ia231ebadbeb9f5575d4b30eff0ce1df6f34178e0