"➕New" Dropdown can display trackers
Run forgeupgrade.
Hack the database and inject one or more tracker id.
On every pages of the project, the "➕New" dropdown displays links to
submit artifact in given trackers.

If user is not allowed to submit an artifact in the tracker, then the
link is not displayed.

The title of the global links has been renamed from Tuleap to
`sys_name`.

Part of story #16205: « + New » button

Change-Id: Ib896b08787c6101e9c850d055d2bf66a47dd649f