Add justification comments in the test plan export
HTML comments are stripped to text because it's the only thing we can do
right now. Putting HTML in a Excel cell means converting it to the rich
text format which is not a trivial task as it requires the capability to
interpret CSS and HTML and is not supported by the xlsx library.

Part of story #15096: export a report of tests for a milestone

Change-Id: I487aaa544fcc6891d48d5614afb0ccb2b7578511