Introduce a way to limit concurrency
While it's not yet that useful thanks to the quite permissive limit of
the REST API endpoint GET /artifacts it will becomes necessary in the
upcoming contributions to not flood the server and keep things
manageable for the client (the tracker representation of each type of
artifact will be retrieved).

To test either build a massive report or play with the constants MAX_CHUNK_SIZE_ARTIFACTS
and MAX_CONCURRENT_REQUESTS_WHEN_RETRIEVING_ARTIFACT_CHUNKS in
artifacts-retriever.ts.

Change-Id: I777339dad11c71546e217d21191f08f4038bf55d