Add 'c' shortcut to open the "➕ New" dropdown
As a user on a given page in a FP or BP page, type `c`. The dropdown "➕
New" is displayed and user can select an entry with the mouse or with
tab/enter combination.

We remove keymaster-sequence (which didn't see a commit since 9 years
and is not in npm) in favor of hotkeys[0].

Note: Help modal and some other shortcuts will be added in dedicated
changesets.

Part of story #16205: « + New » button

[0] https://github.com/jaywcjlove/hotkeys

Change-Id: Ia93f63f97c639e3bbb6342da915da6c5bc9298f7