Use planning backlog tracker SQL table in multi project backlog
This is part of story #16152 [backlog management] disable creation of items in Aggregator projects

We do not rely anymore on a custom SQL table to store and synchronize
contributor backlog trackers. Read only view of top planning in
aggregator project and the synchronization at contributor top planning
update must work as expected.

Change-Id: I651ad18e0ac6bbb1847a26f6511fe92bf057b67d