Do not set a large width on section titles when there are multiple headers
As the column width will be set on the first column you can have a very
large first column when it's not really useful.
Not setting the exact width on the cell containing a section title is
not perfect as you can still have a situation where the title is cut but
it is easy to implement and it is good enough for the vast majority of
the situations (at least it is in en and fr with the current format of
the report).

Part of story #15096: export a report of tests for a milestone

Change-Id: I3fd81703dfb02ccc1639c7a53fbd7ef9ca1bbb9f