Export text and string fields of the requirements in the test plan report
Fields of the semantic title are not added again as the "Title" column is
always present.

Long text content is not really well managed as you can end up with a
very large column. This behavior will be improvec independently as it
also affects the justification comments.

Part of story #15096: export a report of tests for a milestone

Change-Id: Id763971ca1512ec711fc9b2061209ab8b8b89ba8